RSC: Refactor RSC plugin into its own package
