+++ code/lib/rsc-plugin/README.md
@@ -1,0 +1,4 @@
+# RSC Plugin
+
+Experimental plugin to transform React Server Components into
+client components that handle synchronization with the `use` hook.
