+++ code/lib/rsc-plugin/project.json
@@ -1,0 +1,8 @@
+{
+  "name": "rsc-plugin",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "projectType": "library",
+  "targets": {
+    "build": {}
+  }
+}
