+++ code/package.json
@@ -148,6 +148,7 @@
     "@storybook/react": "workspace:*",
     "@storybook/react-vite": "workspace:*",
     "@storybook/react-webpack5": "workspace:*",
+    "@storybook/rsc-plugin": "workspace:*",
     "@storybook/server": "workspace:*",
     "@storybook/server-webpack5": "workspace:*",
     "@storybook/source-loader": "workspace:*",
